+++ template-parts/cover-featured.php
@@ -22,6 +22,9 @@
 		<li class="slider-list-item">
 			<a href="<?php the_permalink(); ?>" rel="bookmark">
 				<div class="cover<?php if ( '' != $img ) { ?> featured-image<?php } ?>">
+					<?php if ( function_exists('has_post_video') && has_post_video() ) { ?>
+						<span class="svg-icon"><?php echo file_get_contents( get_template_directory_uri() . '/dist/svg/play-circle.svg' ); ?></span>
+					<?php } ?>
 					<div class="cover-background"<?php if ( '' != $img ) { ?> style="background-image: url('<?php echo $img; ?>');"<?php } ?>></div>
 					<header class="cover-header">
 						<h2 class="cover-subtitle">Featured</h2>