+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file. This projec
 - Added support for [Yet Another Related Posts Plugin](https://wordpress.org/plugins/yet-another-related-posts-plugin/).
 - Added custom YARPP templates: two column grid, three column grid, slider, full-width slider, bullet list, and comma separated.
 - Added support for Aesop Story Engine hero gallery.
+- Added support for [Featured Video Plus](https://wordpress.org/plugins/featured-video-plus/) plugin.
 - Fixed slider bug where longer titles were being cut off.
 - Fixed slider styles to eliminate page jump.
 - Fixed nested menu bug (thanks Mariusz).